Completo

+++ app/src/main/java/com/example/firestore/MainActivity.kt
@@ -27,9 +27,14 @@ class MainActivity : AppCompatActivity() {
             val editNomeAlt : EditText = findViewById(R.id.Change)
             val nomeAlterar = editNomeAlt.text.toString()
 
-            val i = Intent(this, Alter::class.java)
-            i.putExtra("Nome", nomeAlterar)
-            startActivity(i)
+            if (nomeAlterar.length != 0){
+                val i = Intent(this, Alter::class.java)
+                i.putExtra("Nome", nomeAlterar)
+                startActivity(i)
+            }
+            else{
+                Toast.makeText(this, "Campo nulo inválido", Toast.LENGTH_SHORT).show()
+            }
         }
 
         btnCadastrar.setOnClickListener {
@@ -41,13 +46,18 @@ class MainActivity : AppCompatActivity() {
             val editNomeExcluir : EditText = findViewById(R.id.Change)
             val nomeExcluir = editNomeExcluir.text.toString()
 
-            db.collection("cadastro").document(nomeExcluir)
-                .delete()
-                .addOnSuccessListener { Toast.makeText(this, "Cadastro excluído com sucesso!", Toast.LENGTH_SHORT).show()}
-                .addOnFailureListener { Toast.makeText(this, "Falhou :(", Toast.LENGTH_SHORT).show() }
+            if(nomeExcluir.length != 0){
+                db.collection("cadastro").document(nomeExcluir)
+                    .delete()
+                    .addOnSuccessListener { Toast.makeText(this, "Cadastro excluído com sucesso!", Toast.LENGTH_SHORT).show()}
+                    .addOnFailureListener { Toast.makeText(this, "Falhou :(", Toast.LENGTH_SHORT).show() }
 
-            val i = Intent(this, MainActivity::class.java)
-            startActivity(i)
+                val i = Intent(this, MainActivity::class.java)
+                startActivity(i)
+            }
+            else{
+                Toast.makeText(this, "Campo nulo inválido", Toast.LENGTH_SHORT).show()
+            }
         }
 
         val txt: TextView = findViewById(R.id.textView)
